requirements.txt
absl-py==0.9.0
affine==2.3.0
asn1crypto==0.24.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cntk-gpu==2.7
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Fiona==1.8.13.post1
gast==0.2.2
geopandas==0.7.0
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.6
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.1.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
pyproj==2.6.0
pyrsistent==0.16.0
python-apt==1.6.4
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
rasterio==1.1.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.13.0
sklearn-pandas==1.8.0
snuggs==1.4.7
tensorboard==2.1.1
tensorflow-estimator==2.1.1
tensorflow-gpu==2.1.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
